// sanity.types.ts
export type ExperiencesQueryResult = Array<{
  _id: string;
  role: string | null;
  company: string | null;
  companyLink: string | null;
  year: string | null;
  description: string | null;
  technologies: Array<string> | null;
}>;